level9/requirement.txt
amqp==2.6.1
asgiref==3.5.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==4.4.7
coverage==6.3.2
Deprecated==1.2.13
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
kombu==4.6.11
packaging==21.3
pyparsing==3.0.7
pytz==2021.3
redis==4.1.4
sqlparse==0.4.2
vine==1.3.0
wrapt==1.13.3
